requirements.txt
flask==3.0.3
flask-cors==4.0.1
authlib==1.3.1
pyotp==2.9.0
qrcode[pil]==7.4.2
pandas==2.2.2
azure-storage-blob==12.20.0
gunicorn==22.0.0
requests==2.32.3
python-dotenv==1.0.1